parser: don't format runtime errors as test failures.
The test failure format was originally added in a time when UBSan did
not fail tests, and junit reports were generated by xcpretty. Failing
the test cosmetically in xcpretty was the indication that there was
a runtime error but the tests itself still passed.

This commit fixes the situation by printing the runtime error itself.
Failing the test is not the business of xcpretty.