parser: don't format runtime error info.
The runtime info contained extra information about the runtime error.
For ASAN and TSAN this is the crash dump which contains lots of data
like the crash dump, and it's debatable whether we want to print it or
not.

For UBSAN, there's no real info, since is doesn't contain any crash
information.

This PR keeps the runtime error info, but specifically doesn't format
it.