parser: fix parsing process product packaging.
The regex that was used before was too specific (it only matched
entitlements, whereas sometimes this log also appeared with provisioning
profiles, and also only matched if they were wrapped in `"`, which
wasn't always the case). It also didn't properly work with escaped
spaces in the paths. Tests were added for the above cases and the
parser was edited to pass them.